[REFACTORING] Split IMAPServerSearchTest
